Extract `DirectoryCleaner`
