+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/gradle/compiler/ModelCompilerPlugin.java
@@ -23,7 +23,6 @@ import io.spine.gradle.compiler.lookup.proto.ProtoToJavaMapperPlugin;
 import io.spine.gradle.compiler.validate.ValidatingBuilderGenPlugin;
 import io.spine.tools.gradle.SpinePlugin;
 import io.spine.tools.gradle.compiler.annotation.ProtoAnnotatorPlugin;
-import io.spine.tools.gradle.compiler.cleaning.CleaningPlugin;
 import io.spine.tools.gradle.compiler.enrichment.EnrichmentLookupPlugin;
 import io.spine.tools.gradle.compiler.protoc.ProtocPluginImporter;
 import io.spine.tools.gradle.compiler.rejection.RejectionGenPlugin;
